Linear Search in Array
Added a cpp program, Linear Search in Array:

Declare an integer array of 10 elements and take all inputs values from the user. Further, take a random number from user and search the entire array if the entered number is present in the array. If present, return the array index at which the number is first time found. Otherwise, print "The number is not present".